fixed wrong animation duration values for approach circles and fade in animations, fixed sliders having wrong timing point assigned to them when timing point spawn time was not integer, fixed slider tick showing incorrectly when spawning slider or hitting reverse arrow while seeking backwards, slider ticks now dont give misses when seeking backwards by frame and hitting reverse arrow or using song slider seeking backwards, fixed hit markers spawning incorrectly when seeking backwards

+++ OsuFileParsers/Decoders/BeatmapDecoder.cs
@@ -764,11 +764,11 @@ namespace OsuFileParsers.Decoders
             {
                 int mid = l + ((r - l) >> 1);
 
-                if (time >= timingPoints[mid].Time)
+                if (time >= (int)timingPoints[mid].Time)
                 {
                     l = mid + 1;
                 }
-                else if (time < timingPoints[mid].Time)
+                else if (time < (int)timingPoints[mid].Time)
                 {
                     r = mid;
                 }
